dev-manager/control-pane/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "remram-dev-manager-control-pane"
version = "0.2.0"
description = "Canonical Remram control plane CLI, service runtime, and MCP adapter."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6,<7",
  "fastapi>=0.115,<1",
  "uvicorn>=0.30,<1",
  "mcp[cli]>=1.0,<2",
]

[project.scripts]
remram = "remram_dev_manager_control_pane.__main__:main"
remram-mcp = "remram_dev_manager_control_pane.mcp_server:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
